+++ apps/performance-tests/src/tree-widget/ModelsTree.test.ts
@@ -123,6 +123,7 @@ describe("models tree", () => {
       // Add one element to always draw set to trigger additional queries
       viewport.setAlwaysDrawn(new Set([elementsModel]));
       await handler.changeVisibility(createModelHierarchyNode(elementsModel), true);
+      viewport.renderFrame();
       await validateHierarchyVisibility({
         ignoreChildren: (node) => !!node.extendedData?.isCategory,
         provider,
@@ -287,6 +288,7 @@ describe("models tree", () => {
       const elementsCategory = visibilityTargets.categories[0];
 
       await handler.changeVisibility(createModelHierarchyNode(elementsModel), true);
+      viewport.renderFrame();
       await validateHierarchyVisibility({
         provider,
         handler,
@@ -348,7 +350,6 @@ describe("models tree", () => {
       expect(visibilityTargets.categories.length).to.be.eq(1);
       const elementsCategory = visibilityTargets.categories[0];
       const node = { modelId: elementsModel, elementId: visibilityTargets.elements[0], categoryId: elementsCategory, subjectId: "0x1" };
-
       await validateHierarchyVisibility({
         provider,
         handler,
@@ -367,6 +368,7 @@ describe("models tree", () => {
     },
     test: async ({ viewport, handler, provider, node }) => {
       await handler.changeVisibility(createElementHierarchyNode(node), true);
+      viewport.renderFrame();
       await validateHierarchyVisibility({
         provider,
         handler,