+++ packages/itwin/measure-tools/src/measurements/AreaMeasurement.ts
@@ -172,7 +172,6 @@ export class AreaMeasurement extends Measurement {
 
     if (props) this.readFromJSON(props);
 
-    this.populateFormattingSpecsRegistry().catch();
   }
 
   private handleTextMarkerButtonEvent(ev: BeButtonEvent): boolean {