+++ packages/itwin/tree-widget/src/tree-widget-react/components/trees/external-sources-tree/ExternalSourcesTreeDefinition.ts
@@ -85,6 +85,8 @@ export class ExternalSourcesTreeDefinition implements HierarchyDefinition {
       filter: props.instanceFilter,
       contentClass: { fullName: "BisCore.ExternalSource", alias: "this" },
     });
+
+    // cspell:disable
     return [
       {
         fullClassName: "BisCore.ExternalSource",
@@ -112,6 +114,7 @@ export class ExternalSourcesTreeDefinition implements HierarchyDefinition {
         },
       },
     ];
+    // cspell:enable
   }
 
   private async createExternalSourcesGroupChildrenQuery({
@@ -122,6 +125,7 @@ export class ExternalSourcesTreeDefinition implements HierarchyDefinition {
       filter: instanceFilter,
       contentClass: { fullName: "BisCore.ExternalSource", alias: "this" },
     });
+    // cspell:disable
     return [
       {
         fullClassName: "BisCore.ExternalSource",
@@ -151,6 +155,7 @@ export class ExternalSourcesTreeDefinition implements HierarchyDefinition {
         },
       },
     ];
+    // cspell:enable
   }
 
   private async createExternalSourceChildrenQuery({